chore(serviceaccounts): Do not update service account if no change in roles
After the introduction of the OR mode for checking permissions, a
regular user should be able to modify the pipeline if it has any of the
roles in the pipeline (not necessary all of them). However, the service
user is created in every save operation, which prevents users to update
the pipeline when the OR mode is enabled.

This patch skips creating/updating the service user in case the user
already existed and the roles were the same as in the pipeline
definition. This change allows the user to udpate the pipeline only when
the roles are not changed to avoid privilege scalation.